# agenda-consultas/swagger-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agenda Consultas API",
    "description": "Versao anterior da API REST para gerenciamento de consultas medicas.",
    "contact": {
      "name": "Suporte",
      "email": "suporte@agendaconsultas.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra um novo usuario e retorna o token JWT.",
        "description": "Exemplo:\r\n{\r\n  \"email\": \"usuario@email.com\",\r\n  \"password\": \"123456\",\r\n  \"role\": \"usuario\",\r\n  \"nome\": \"Ana Silva\",\r\n  \"cpf\": \"12345678901\",\r\n  \"telefone\": \"11999999999\"\r\n}",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica um usuario e retorna o token JWT.",
        "description": "Exemplo:\r\n{\r\n  \"email\": \"admin@email.com\",\r\n  \"password\": \"123456\"\r\n}",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/role": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Atualiza o role de um usuario (admin).",
        "description": "Exemplo:\r\n{\r\n  \"email\": \"usuario@email.com\",\r\n  \"role\": \"admin\"\r\n}",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRoleUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRoleUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRoleUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRoleUpdateResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRoleUpdateResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRoleUpdateResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consultas/{id}": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Busca uma consulta pelo id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consulta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Consultas"
        ],
        "summary": "Atualiza uma consulta por id (admin).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Consultas"
        ],
        "summary": "Atualiza parcialmente uma consulta por id (admin).",
        "description": "Envie apenas os campos que deseja alterar. Exemplo:\r\n{\r\n  \"status\": \"concluida\"\r\n}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaPatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaPatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaPatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Consultas"
        ],
        "summary": "Remove uma consulta por id (admin).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/consultas/paciente/{pacienteId}": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Lista consultas por paciente.",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/consultas/slots": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Lista horarios padronizados e disponibilidade por dia.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaSlotDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaSlotDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaSlotDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthLoginDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AuthRegisterDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthRoleUpdateDto": {
        "required": [
          "email",
          "role"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "role": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AuthRoleUpdateResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Consulta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "pacienteId": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "dataBrasil": {
            "type": "string",
            "nullable": true
          },
          "especialidade": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ConsultaPatchDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "especialidade": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConsultaSlotDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "disponivel": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ConsultaUpdateDto": {
        "required": [
          "data",
          "especialidade",
          "pacienteId"
        ],
        "type": "object",
        "properties": {
          "pacienteId": {
            "minLength": 1,
            "type": "string"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "especialidade": {
            "minLength": 3,
            "type": "string"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Informe o token JWT no formato: Bearer {seu_token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}